% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bunzip2.R
\name{bunzip2}
\alias{bunzip2}
\title{Uncompress a file with bunzip2}
\usage{
bunzip2(f, keep = TRUE)
}
\arguments{
\item{f}{The file to uncompress}

\item{keep}{Whether to keep the compressed file}
}
\value{
Path to the uncompressed file
}
\description{
Uncompress a file with bunzip2
}

