% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimpipeline-package.R
\docType{package}
\name{jimpipeline-package}
\alias{jimpipeline}
\alias{jimpipeline-package}
\title{Package to wrap Jefferis lab image processing pipeline}
\description{
This will contain functions that we use during our large scale image 
processing and registration work. We will try to make these as generic as
possible, but for the moment use by third parties is not a major design
consideration.
}
\keyword{package}

