% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_metadata.R
\name{lsm_metadata}
\alias{lsm_metadata}
\title{Extract the metadata from Zeiss LSM file using loci bioformats as text vector}
\usage{
lsm_metadata(
  f,
  cachefile = TRUE,
  ReturnMetaData = TRUE,
  Force = FALSE,
  UseLock = FALSE
)
}
\arguments{
\item{f}{Path to lsm file}

\item{cachefile}{Whether to save a copy of metadata to disk (TRUE)}

\item{ReturnMetaData}{Whether to return metadata rather than success (default:
\code{TRUE})}

\item{Force}{whether to re-parse metadata even if a cached version exists}

\item{UseLock}{whether to use a lock file while parsing for simple parallelism}
}
\value{
character vector of metadata OR TRUE/FALSE for success
}
\description{
Specifically, this relies on the showinf tool of loci Details
}
\examples{
\dontrun{
lsmdir=file.path(fcconfig$regroot,"lsms")
for(f in dir(lsmdir,patt="lsm$",full=T)) lsm_metadata(f,UseLock=TRUE)
}
}
\seealso{
\code{\link{parse_key_lsm_metadata}}
}
