% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_metadata.R
\name{make_lsm_df}
\alias{make_lsm_df}
\title{Make summary dataframe for lsm files and their key metadata}
\usage{
make_lsm_df(lsmdir, oldlsmdf = NULL, extrafields = NULL, Verbose = TRUE)
}
\arguments{
\item{lsmdir}{Path to directory containing lsm files}

\item{oldlsmdf}{Dataframe or path to rda file containing one}

\item{extrafields}{Function to parse extra fields from metadata for each image}

\item{Verbose}{Logical: Show messages about progress/updating of cached data}
}
\value{
dataframe
}
\description{
Will calculate md5 sums for all input files and then use this to update a
  dataframe.
If an extrafields function is supplied, it should take one parameter, the
  list output of parse_key_lsm_metadata(), which will include as an
  attribute the full contents of the metadata file dumped by the loci
  showinf tool. The function should take care to return values for all
  fields that it _could_ return each time it is called even if these are NA
  and to ensure that values have a consistent data type per field.
}
\seealso{
\code{\link{lsm_metadata}}
}
