% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_metadata.R
\name{parse_key_lsm_metadata}
\alias{parse_key_lsm_metadata}
\title{Parse key Zeiss LSM metadata into an R list}
\usage{
parse_key_lsm_metadata(f, text = NULL, ReturnRawMetaData = FALSE)
}
\arguments{
\item{f}{Path to file containing lsm metadata}

\item{text}{Text version of lsm metadata (optional, otherwise read from file)}

\item{ReturnRawMetaData}{Whether to return the raw metadata from the file.}
}
\value{
a list containing parsed metadata
}
\description{
Parse key Zeiss LSM metadata into an R list
}
\seealso{
\code{\link{lsm_metadata}}
}
