% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickhash.R
\name{quickhash.lsm}
\alias{quickhash.lsm}
\title{Compute a quick hash for head and tail of a Zeiss LSM confocal file}
\usage{
quickhash.lsm(f, offsets = c(0, -1e+05), chunksizes = c(150000, 1e+05), ...)
}
\arguments{
\item{f}{Path to lsm files}

\item{offsets}{Default c(0,-1e5)}

\item{chunksizes}{Default c(1e5,1e5)}

\item{...}{Additional arguments passed to default \code{\link{quickhash}} function}
}
\value{
character vector containing hash values for lsm(s)
}
\description{
The first ~ 150,000 bytes includes a number of fields such as dimension and
acquisition timestamp that are typically enough to make for a unique hash. In
particular the first 150K (but not 100K) are unique across the 16226 v2
flycircuit LSMs. The hope is the last 100K should include some image data
(though in practice for flycircuit LSMs) this is often in the 8 bit
segmented neurons which nearly all have 0 background
}
\seealso{
\code{\link{quickhash}} and \code{\link{digest}}
}
