% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji.R
\name{runFijiMacro}
\alias{runFijiMacro}
\title{Run a Fiji macro}
\usage{
runFijiMacro(macro = "", macroArg = "", headless = FALSE, batch = TRUE,
  MinMem = MaxMem, MaxMem = "2500m", IncrementalGC = TRUE,
  Threads = NULL, fijiArgs = NULL, javaArgs = NULL, ijArgs = NULL,
  fijiPath = fiji(), DryRun = FALSE)
}
\arguments{
\item{macro}{Path to macro to run}

\item{macroArg}{Arguments for macro}

\item{headless}{Whether to have ImageJ/Fiji running headless}

\item{batch}{Use batch mode}

\item{MinMem, MaxMem}{Memory limits}

\item{IncrementalGC}{Whether to use incremental garbage collection}

\item{Threads}{Number of threads}

\item{fijiArgs}{Arguments for ImageJ/Fiji (as opposed to the macro itself)}

\item{javaArgs}{Arguments for java}

\item{ijArgs}{Arguments for ImageJ}

\item{fijiPath}{Path to fiji executable (can be set by
\code{options(jimpipeline.fiji="/some/path")})}

\item{DryRun}{Whether to return the command to be run rather than actually 
executing it.}
}
\description{
Run a Fiji macro
}

