% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skel.R
\name{skeletonise_nrrds}
\alias{skeletonise_nrrds}
\title{Skeletonise NRRD images (e.g. to make them suitable for NBLAST)}
\usage{
skeletonise_nrrds(input, outdir, threshold = 0, fileregex = NULL,
  DryRun = TRUE, ...)
}
\arguments{
\item{input}{Path to single input file or an input directory}

\item{outdir}{Path to output directory}

\item{threshold}{Integer specifying an absolute threshold or string
specifying ImageJ authothreshold method.}

\item{fileregex}{Optional regex matching the \bold{whole} of the filename (NB
just filename without parent directory)}

\item{DryRun}{Show what would happen but don't run (default TRUE)}

\item{...}{Additional arguments passed to \code{\link{runFijiMacro}}}
}
\value{
Logical indicating if the command completed successfully
}
\description{
Skeletonise NRRD images (e.g. to make them suitable for NBLAST)
}
\examples{
\dontrun{
# basic version
skeletonise_nrrds('/path/to/indir', '/path/to/outdir')
# convert those to dotrops for NBLAST
dps=dotprops(dir('/path/to/outdir', pattern='nrrd$', full.names=TRUE))

# additional arguments
skeletonise_nrrds('/path/to/indir', '/path/to/outdir', threshold='Iso_Data')
skeletonise_nrrds('/path/to/indir', '/path/to/outdir', fileregex='^seg_whole.*c0\\\\.nrrd')
}
}
