% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimpipeline-package.R
\docType{package}
\name{jimpipeline-package}
\alias{jimpipeline}
\alias{jimpipeline-package}
\title{Package to wrap Jefferis lab image processing pipeline}
\description{
This contains functions that we use during our large scale image processing 
and registration work. We will try to make these as generic as possible, but 
for the moment use by third parties is not a major design consideration.
}
\section{Package Options}{
 The following options can be set to specify default
  behaviour.
  
  \itemize{
  
  \item{\code{jimpipeline.fiji}}{ Location of fiji to use for running 
  scripts}
  
  If this is not set, at the moment the package will check the path for an 
  executable called fiji.
  
  }
}
\keyword{package}

