% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quickhash}
\alias{quickhash}
\title{Compute a (quick) hash for some part of the contents of a set of files}
\usage{
quickhash(f, offsets = 0, chunksizes = Inf, CheckFileSize = TRUE, ...)
}
\arguments{
\item{f}{Path to files to hash}

\item{offsets}{integer offsets into file at which to read data}

\item{chunksizes}{amount of data to read at each offset}
}
\value{
character vector containing hash values for file(s)
}
\description{
Default is to compute md5 hash for whole of each file.
negative offsets => seek to offset position wrt end of file
chunksize=Inf => whole file
Uses digest() which defaults to algo=md5
}
\seealso{
\code{\link{digest}}
}

