% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiji.R
\name{runFijiMacro}
\alias{runFijiMacro}
\alias{fiji}
\title{Call external Fiji/ImageJ executable}
\usage{
runFijiMacro(macro = "", macroArg = "", headless = FALSE,
  batch = TRUE, MinMem = MaxMem, MaxMem = "2500m",
  IncrementalGC = TRUE, Threads = NULL, fijiArgs = NULL,
  javaArgs = NULL, ijArgs = NULL, fijiPath = fiji(),
  DryRun = FALSE)

fiji(fijiPath = NULL)
}
\arguments{
\item{macro}{Path to macro to run}

\item{macroArg}{Arguments for macro}

\item{headless}{Whether to have ImageJ/Fiji running headless}

\item{batch}{Use batch mode}

\item{MinMem, MaxMem}{Memory limits}

\item{IncrementalGC}{Whether to use incremental garbage collection}

\item{Threads}{Number of threads}

\item{fijiArgs}{Arguments for ImageJ/Fiji (as opposed to the macro itself)}

\item{javaArgs}{Arguments for java}

\item{ijArgs}{Arguments for ImageJ}

\item{fijiPath}{Path to fiji executable (can be set by
\code{options(jimpipeline.fiji="/some/path")})}

\item{DryRun}{Whether to return the command to be run rather than actually 
executing it.}
}
\value{
logical indicating if the command completed successfully
}
\description{
\code{runFijiMacro} runs a macro using preferred Fiji executable

\code{fiji} returns path to preferred Fiji executable
}
\examples{
# Path to current Fiji executable
\donttest{
fiji()
}

\dontrun{
# Set path to preferred Fiji executable (this will be remembered)
# you can also set options(jimpipeline.fiji="/some/path")
fiji("/Applications/Fiji.app/Contents/MacOS/ImageJ-macosx")
}
}
